some changes

+++ code/run_server.sh
@@ -1,4 +1,6 @@
 #echo make sure you have python2.7 + django in your environment
+
+
 if [ -z $ESGF_NM_TASKS ]; then
     echo please set ESGF_NM_TASKS to a directory for temporary task storage
     exit 1
@@ -11,6 +13,13 @@ if [ -z $ESGF_NODEMGR_MAP ] ; then
     exit 1
 fi
 
+if [ -z $ESGF_NM_PORT ] ; then
+    echo please set ESGF_NM_PORT
+    exit 1
+fi
+
+
+
 python server/basic_nodemanager.py $ESGF_NODEMGR_MAP $ESGF_NM_TIMESTAMP &
-server/nodemgr/manage.py runserver 0.0.0.0:80
+python server/nodemgr/manage.py runserver 0.0.0.0:$ESGF_NM_PORT
 kill %1
